# packages/kafka-event-consumer/test/config/test.swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "/test",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://test.foo"
        }
    ],
    "paths": {
...
        "/test/bulk": {
            "get": {
                "summary": "Retrieves a list of all documents",
                "operationId": "bulk_read",
                "responses": {
                    "200": {
                        "description": "The documents were successfully retrieved.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/State"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Creates multiple documents",
                "operationId": "bulk_create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/State"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The documents were successfully created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "successful": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OperationStatus"
                                            }
                                        },
                                        "failed": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OperationStatus"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Deletes multiple documents",
                "operationId": "bulk_delete",
                "responses": {
                    "200": {
                        "description": "The documents were successfully deleted.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "successful": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OperationStatus"
                                            }
                                        },
                                        "failed": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OperationStatus"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "State": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "format": "id"
                    },
                    "name": {
                        "type": "string",
                        "faker": "animal.horse"
                    },
                    "eggs": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 10
                    },
                    "coop_id": {
                        "type": "string",
                        "format": "id"
                    }
                },
                "required": [
                    "name"
                ],
                "additionalProperties": false
            },
            "OperationStatus": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "error": {
                        "type": "string"
                    }
                }
            }
        }
    }
}